# Conflicts:
#	.gitlab-ci.yml

+++ modules/support.js
@@ -10,7 +10,7 @@ module.exports = {
     name: "Support Server Specific Functions",
     init: function(bot){
         bot.client.on("message", async function onMessage(message) {
-            if (message.guild && message.guild.id === "322032568558026753" && !message.author.bot) {
+            if (message.guild && message.guild.id === "322032568558026753" && !message.author.bot && !bot.client.user.id == "146293573422284800") {
                 if (message.content.indexOf("discord.gg") > -1)
                     return message.delete();
